+++ lib/src/search/search_client.dart
@@ -1,8 +1,11 @@
+import 'package:youtube_explode_dart/src/reverse_engineering/responses/search_page.dart';
+
 import '../common/common.dart';
 import '../reverse_engineering/responses/playlist_response.dart';
 import '../reverse_engineering/youtube_http_client.dart';
 import '../videos/video.dart';
 import '../videos/video_id.dart';
+import 'base_search_content.dart';
 import 'search_query.dart';
 
 /// YouTube search queries.
@@ -13,7 +16,8 @@ class SearchClient {
   SearchClient(this._httpClient);
 
   /// Enumerates videos returned by the specified search query.
-  Stream<Video> getVideosAsync(String searchQuery) async* {
+  /// (from the YouTube Embedded API)
+  Stream<Video> getVideos(String searchQuery) async* {
     var encounteredVideoIds = <String>{};
 
     for (var page = 0; page < double.maxFinite; page++) {
@@ -49,7 +53,41 @@ class SearchClient {
     }
   }
 
+  /// Enumerates videos returned by the specified search query
+  /// (from the video search page).
+  /// Contains only instances of [SearchVideo] or [SearchPlaylist]
+  Stream<BaseSearchContent> getVideosFromPage(String searchQuery) async* {
+    var page = await SearchPage.get(_httpClient, searchQuery);
+    yield* Stream.fromIterable(page.initialData.searchContent);
+
+    // ignore: literal_only_boolean_expressions
+    while (true) {
+      page = await page.nextPage(_httpClient);
+      if (page == null) {
+        return;
+      }
+      yield* Stream.fromIterable(page.initialData.searchContent);
+    }
+  }
+
   /// Queries to YouTube to get the results.
+  @Deprecated('Use getVideosFromPage instead')
   Future<SearchQuery> queryFromPage(String searchQuery) =>
       SearchQuery.search(_httpClient, searchQuery);
 }
+
+/*
+    channelId = ChannelId.fromString(channelId);
+    var page = await ChannelUploadPage.get(
+        _httpClient, channelId.value, videoSorting.code);
+    yield* Stream.fromIterable(page.initialData.uploads);
+
+    // ignore: literal_only_boolean_expressions
+    while (true) {
+      page = await page.nextPage(_httpClient);
+      if (page == null) {
+        return;
+      }
+      yield* Stream.fromIterable(page.initialData.uploads);
+    }
+ */