+++ CHANGELOG.md
@@ -3,9 +3,15 @@
 - Only throw custom exceptions from the library.
 - `getUploadsFromPage` no longer throws.
 
+## 1.5.1
+- BREAKING CHANGE: Renamed `getVideosAsync` to `getVideos`.
+- Implemented `getVideosFromPage` which supersedes `queryFromPage`.
+- Implemented JSON Classes for reverse engineer.
+- Added `forceWatchPage` to the video client to assure the fetching of the video page. (ATM useful only if using the comments api)
+- Remove adaptive streams. These are not used anymore. 
+ 
 ## 1.5.0
 - BREAKING CHANGE: Renamed `Container` class to `StreamContainer` to avoid conflicting with Flutter `Container`. See #66
-
 
 ## 1.4.4
 - Expose HttpClient in APIs